Finalize modbus master and add good test script
